Iván-Solucionado verificación de campos al añadir trainer, añadida las funcionalidades de que el admin pueda ver todos los usuarios desde su perfil y eliminar un comentario específico desde la sección de comentarios.

+++ gymapp/src/main/java/es/codeurjc/gymapp/services/CommentService.java
@@ -26,9 +26,15 @@ public class CommentService {
         commentsRepository.save(comment);
     }
 
-    public void delete(Long commentId, Trainer trainer) {
-        Comment comment = this.findById(commentId).get();
-        trainer.getComments().remove(comment);
-        commentsRepository.delete(comment);
+    public boolean delete(Long commentId, Trainer trainer) {
+        Optional<Comment> opComment = this.findById(commentId); 
+        Comment comment;
+        if (opComment.isPresent()) {
+            comment = opComment.get();
+            trainer.getComments().remove(comment);
+            commentsRepository.delete(comment);
+            return true;
+        }
+        return false;
     }
 }
